+++ Server.Logic/Abstract/Managers/IShoppingListManager.cs
@@ -9,7 +9,7 @@ public interface IShoppingListManager
     public Task<IManagerActionResult> DeleteShoppingListAsync(int shoppingListId);
     public Task<IManagerActionResult<ShoppingList[]>> GetAllShoppingListsAsync(string user);
     public Task<IManagerActionResult<ShoppingList>> GetShoppingListAsync(int shoppingListId);
-    public Task<IManagerActionResult> UpdateShoppingListAsync(int shoppingListId, int categoryId, int quantity);
+    public Task<IManagerActionResult> UpdateShoppingListAsync(int shoppingListId, int categoryId, double quantity);
     public Task<IManagerActionResult<(int Id, string Name)[]>> GetShoppingListMapAsync(string user);
     public Task<IManagerActionResult<int>> CreateShoppingListByRecipeAsync(int recipeId, int[] fridgesId, string user);
 }
